renaming jooqflyway to jooqjobactivityhistory
