+++ titus-ext/jooq-job-activity-history/gradle/dependency-locks/compileClasspath.lockfile
