+++ titus-ext/jooq-job-activity-history/src/main/java/com/netflix/titus/ext/jobactivityhistory/JooqJobActivityContextComponent.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.titus.ext.jooqflyway;
+package com.netflix.titus.ext.jobactivityhistory;
 
 
 import com.netflix.titus.common.util.archaius2.Archaius2Ext;
@@ -31,12 +31,21 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 import org.springframework.core.env.Environment;
 
+import javax.inject.Named;
+
 @Configuration
 public class JooqJobActivityContextComponent {
 
     @Bean
+    @Named("jobActivityJooqContext")
     public JooqConfiguration getJooqPropertyConfiguration(Environment environment) {
         return Archaius2Ext.newConfiguration(JooqConfiguration.class, "titus.ext.supplementary.jobactivity", environment);
+    }
+
+    @Bean
+    @Named("producerJooqContext")
+    public JooqConfiguration getJooqPropertyProducerConfiguration(Environment environment) {
+        return Archaius2Ext.newConfiguration(JooqConfiguration.class, "titus.ext.supplementary.jobproducer", environment);
     }
 
     @Bean
@@ -47,7 +56,7 @@ public class JooqJobActivityContextComponent {
     @Bean
     @Primary
     @Qualifier("jobActivityJooqContext")
-    public JooqContext getJobActivityJooqContext(JooqConfiguration jooqConfiguration, ConfigurableApplicationContext applicationContext) {
+    public JooqContext getJobActivityJooqContext(@Named("jobActivityJooqContext")JooqConfiguration jooqConfiguration, ConfigurableApplicationContext applicationContext) {
         if (jooqConfiguration.isInMemoryDb()) {
             return new EmbeddedJooqContext(applicationContext, "jobActivityJooqContext");
         }
@@ -56,7 +65,7 @@ public class JooqJobActivityContextComponent {
 
     @Bean
     @Qualifier("producerJooqContext")
-    public JooqContext getJooqProducerContext(JooqConfiguration jooqConfiguration, ConfigurableApplicationContext applicationContext) {
+    public JooqContext getJooqProducerContext(@Named("producerJooqContext")JooqConfiguration jooqConfiguration, ConfigurableApplicationContext applicationContext) {
         if (jooqConfiguration.isInMemoryDb()) {
             return new EmbeddedJooqContext(applicationContext, "producerJooqContext");
         }
