+++ titus-ext/jooq-job-activity-history/src/main/resources/db/migration/jobactivity/V1__initial_setup.sql
